fixed logistic reg implementation
