--- code/sto_grad.py
@@ -1,6 +1,0 @@
-import numpy
-
-def sto_grad(y, tx, w, i_t):
-    y_sgd = y[i_t]
-    x_sgd = np.array(tx[i_t,:])
-    return gradient(y_sgd, x_sgd, w)
